[NEW] Detect Plesk resellers that have a system account with the same username. If the conflicting reseller usernames exist, they'll show up like this:
s:Plesk11:RESELLER_CONFLICT:summit

Then when you run cpeval2 with the input file, it will print a message like this, followed by the standard steps for changing a username:

--
========================================
Reseller Username Conflicts
========================================
Plesk allows duplicate usernames for resellers and system accounts, while cPanel does not.
As such, the following usernames will need to be renamed before they can be copied:

* summit

When renaming the users, please use the following guidelines:
[...]
--

----------------------

[UPDATED] Tidied up the duplicate email username output, so now it will look something like this:

Duplicate email users:
 [subscription1.test] alice, Alice, bob, BOB
 [example.test] carol, Carol

(This is all that exists for this feature at this time, since we have no supported workarounds).

----------------------

[UPDATED] Moved the feature that prints instructions on how to change usernames into its own sub called print_username_change_info().